# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Butwal Hacks Internal API",
    "description": "Complete API reference for the Butwal Hacks platform. Covers all 18 internal API routes.\n\n**Base URL:** `http://localhost:3000` (development) or `https://butwalhacks.com` (production)\n\n**Auth:** Most endpoints require a valid Supabase session. The session token is managed via cookies.\n\n**Methods:** All routes use standard Next.js API route handlers.",
    "version": "1.0.0",
    "contact": {
      "name": "Butwal Hacks Maintainers",
      "url": "https://butwalhacks.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://butwalhacks.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Auth0 session cookie. Set automatically after login via `auth()` from `@auth0/nextjs-auth0`. Most endpoints extract the user ID from the active session."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for development/testing. All auth is session-based via Auth0."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" }
        }
      }
    }
  },
  "security": [
    { "sessionAuth": [] }
  ],
  "paths": {
    "/api/contact": {
      "post": {
        "summary": "Submit a contact form message",
        "description": "Sends an email via Resend to the organization's contact address. No auth required — public form endpoint.",
        "tags": ["Public"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "subject", "message"],
                "properties": {
                  "name": { "type": "string", "minLength": 2, "description": "Sender's full name" },
                  "email": { "type": "string", "format": "email", "description": "Sender's email address" },
                  "phone": { "type": "string", "description": "Optional phone number" },
                  "subject": { "type": "string", "minLength": 5, "description": "Message subject line" },
                  "message": { "type": "string", "minLength": 10, "description": "Message body" }
                }
              },
              "example": {
                "name": "Aarav Sharma",
                "email": "aarav@example.com",
                "subject": "Partnership Inquiry",
                "message": "I'd like to discuss a potential partnership with Butwal Hacks."
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Message sent successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Invalid form data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Email provider error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/events/register": {
      "post": {
        "summary": "Register for an event",
        "description": "Registers the authenticated user for an event. Requires Supabase session cookie.",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["event_id"],
                "properties": {
                  "event_id": { "type": "string", "format": "uuid", "description": "Event ID to register for" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registration successful" },
          "401": { "description": "Unauthorized — not logged in" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/events/checkin": {
      "post": {
        "summary": "Check in an attendee",
        "description": "Marks an event registration as attended. Requires organizer or maintainer role.",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["registration_id"],
                "properties": {
                  "registration_id": { "type": "string", "format": "uuid", "description": "Registration ID to mark as attended" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Check-in successful" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
    "/api/projects/like": {
      "post": {
        "summary": "Toggle project like",
        "description": "Toggles the like state for the authenticated user on a project. Calls `toggle_project_like` RPC function.",
        "tags": ["Projects"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["project_id"],
                "properties": {
                  "project_id": { "type": "string", "format": "uuid", "description": "Project ID to like/unlike" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Like toggled" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/profile/complete": {
      "post": {
        "summary": "Complete profile onboarding",
        "description": "Updates the authenticated user's profile fields during the onboarding flow.",
        "tags": ["Profile"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": { "type": "string" },
                  "bio": { "type": "string" },
                  "avatar_url": { "type": "string", "format": "uri" },
                  "github_username": { "type": "string" },
                  "skills": { "type": "array", "items": { "type": "string" } },
                  "socials": {
                    "type": "object",
                    "properties": {
                      "github": { "type": "string" },
                      "linkedin": { "type": "string" },
                      "twitter": { "type": "string" },
                      "website": { "type": "string" }
                    }
                  }
                }
              },
              "example": {
                "full_name": "Jane Hacker",
                "bio": "Full-stack developer passionate about open source.",
                "skills": ["React", "TypeScript", "Node.js"],
                "github_username": "janehacker"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/profile/update": {
      "post": {
        "summary": "Update profile fields",
        "description": "Updates the authenticated user's profile fields. Same as `/api/profile/complete` but can be used for partial updates after onboarding.",
        "tags": ["Profile"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bio": { "type": "string" },
                  "avatar_url": { "type": "string", "format": "uri" },
                  "skills": { "type": "array", "items": { "type": "string" } },
                  "socials": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  },
...
}